+++ src/quicksort.hpp
@@ -5,8 +5,8 @@
 #include <string>
 using namespace std;
 
-void ordena(string tipo, int esquerda, int direita, int *vetor, int originaltamanho, unsigned long long int *numeroComparacoes, unsigned long long int *numeroTrocas);
-void particao(string variacaoQuicksort, int esquerda, int direita, int *i, int *j, int *vetor, unsigned long long int *numeroComparacoes, unsigned long long int *numeroTrocas);
-void insercao(int *vetor, int esquerda, int direita, unsigned long long int *numeroComparacoes, unsigned long long int *numeroTrocas);
+void ordena(string tipo, int esquerda, int direita, int *vetor, int originaltamanho, double *numeroComparacoes, double *numeroTrocas);
+void particao(string variacaoQuicksort, int esquerda, int direita, int *i, int *j, int *vetor, double *numeroComparacoes, double *numeroTrocas);
+void insercao(int *vetor, int esquerda, int direita, double *numeroComparacoes, double *numeroTrocas);
 
 #endif
